#313 Add generateRandomStringWithoutSimilarChars to front-end side
