+++ lib/utils.js
@@ -6,44 +6,9 @@ const path = require('path');
 const { SHA3 } = require('sha3');
 const logs = require('./logs');
 const validateCodes = require('./validateCodes');
+const random = require('../vendors/random');
 
 const { DATABASE_DATA_ERROR, DATABASE_NO_RESULT_ERROR } = validateCodes;
-const allCharacters = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
-const regexForUnreadableChars = /[O0Il]/g;
-const readableCharacters = allCharacters.replace(regexForUnreadableChars, '');
-
-/**
- * @description Generate random string based on given length and own
- * set of characters if you want
- * @param length {number}
- * @param ownCharacters {string}
- * @return {string}
- */
-function generateRandomString (length, ownCharacters = undefined) {
-  let result = '';
-  const characters = ownCharacters || allCharacters;
-  const charactersLength = characters.length;
-  for (let i = 0; i < length; i++) {
-    result += characters.charAt(Math.floor(Math.random() * charactersLength));
-  }
-  return result;
-}
-/**
- * @description Generate random easy to read string based on given length
- * @param length {number}
- * @return {string}
- */
-function generateRandomStringWithoutSimilarChars (length) {
-  return generateRandomString(length, readableCharacters);
-}
-/**
- * @description Check if string is easy to read
- * @param text {string}
- * @return {boolean}
- */
-function checkIfStringIsReadable (text) {
-  return text.match(regexForUnreadableChars) === null;
-}
 
 /**
  * @description Validate config file
@@ -78,7 +43,7 @@ function getSHA (text) {
  * @return {string}
  */
 function generateAccessKey () {
-  return getSHA(generateRandomString(10));
+  return getSHA(random.generateRandomString(10));
 }
 
 /**
@@ -152,11 +117,8 @@ function responseError (response, httpStatus, errorMsg, responseObject) {
 }
 
 module.exports = {
-  checkIfStringIsReadable,
   getSystemConfig,
   getSHA,
-  generateRandomString,
-  generateRandomStringWithoutSimilarChars,
   generateAccessKey,
   envIsDevelopment,
   throwIf,
@@ -165,4 +127,7 @@ module.exports = {
   responseDatabaseNoData,
   responseUserError,
   responseError,
+  checkIfStringIsReadable: random.checkIfStringIsReadable,
+  generateRandomString: random.generateRandomString,
+  generateRandomStringWithoutSimilarChars: random.generateRandomStringWithoutSimilarChars,
 };
